remove debugg prints

+++ obsplus/bank/core.py
@@ -3,9 +3,8 @@ Bank ABC
 """
 import gc
 import os
-import time
-import tables
 import threading
+import time
 import warnings
 from abc import ABC, abstractmethod
 from contextlib import contextmanager, suppress
@@ -13,12 +12,14 @@ from os.path import join
 from pathlib import Path
 from typing import Optional
 
+import numpy as np
 import pandas as pd
+import tables
 from pandas.io.sql import DatabaseError
 
 import obsplus
-from obsplus.utils import iter_files
 from obsplus.exceptions import BankDoesNotExistError, BankIndexLockError
+from obsplus.utils import iter_files
 
 
 class _Bank(ABC):
@@ -30,16 +31,16 @@ class _Bank(ABC):
     # hdf5 compression defaults
     _complib = "blosc"
     _complevel = 9
-
-    # lock for updating index
-    _index_lock = threading.RLock()
     # attributes subclasses need to define
     ext = ""
     bank_path = ""
     namespace = ""
     index_name = ".index.h5"  # name of index file
+    # concurrency handling features
     _lock_file_name = ".~obsplus_hdf5.lock"
     _owns_lock = False
+    _concurrent = False
+    _index_lock = threading.RLock()  # lock for updating index thread
     # optional str defining the directory structure and file name schemes
     path_structure = None
     name_structure = None
@@ -178,21 +179,26 @@ class _Bank(ABC):
         max_retry
             The number of times to retry before raising.
         """
+        # if the concurrent feature is not selected just bail out
+        if not self._concurrent:
+            return
         # if there is no lock, or this bank owns the lock return
         if not os.path.exists(self.lock_file_path) or self._owns_lock:
             return
-        # else try to wait {wait_interval} seconds {max_retry} times
+        # get random state, based on process id, for waiting randomw seconds
+        rand = np.random.RandomState(os.getpid() or 0)
+        # wait until the lock file is gone or timeout occurs (then raise)
         count = 0
         while count < max_retry:
             if not os.path.exists(self.lock_file_path):
                 return
-            time.sleep(wait_interval)
+            time.sleep(wait_interval + rand.rand() * wait_interval)
             count += 1
         else:
             duration = max_retry * wait_interval
             msg = (
-                f"{self.lock_file_path} was not released after "
-                f"{duration} seconds. It may need to be manually deleted"
+                f"{self.lock_file_path} was not released after about "
+                f"{duration * 1.5} seconds. It may need to be manually deleted"
                 f" if the index is not in the process of being updated."
             )
             raise BankIndexLockError(msg)
@@ -202,14 +208,16 @@ class _Bank(ABC):
         """
         Acquire lock for work inside context manager.
         """
-        self.block_on_index_lock()  # ensure lock isn't already in use
-        assert Path(self.bank_path).exists()
-        open(self.lock_file_path, "w").close()  # create lock file
-        self._owns_lock = True
-        # close all open files (nothing should have tables at this point)
-        gc.collect()  # must call GC to ensure everything is cleaned up (ugly)
-        tables.file._open_files.close_all()
+        if self._concurrent:
+            self.block_on_index_lock()  # ensure lock isn't already in use
+            assert Path(self.bank_path).exists()
+            open(self.lock_file_path, "w").close()  # create lock file
+            self._owns_lock = True
+            # close all open files (nothing should have tables at this point)
+            gc.collect()  # must call GC to ensure everything is cleaned up (ugly)
+            tables.file._open_files.close_all()
         yield
-        with suppress(FileNotFoundError):
-            os.unlink(self.lock_file_path)
-        self._owns_lock = False
+        if self._concurrent:
+            with suppress(FileNotFoundError):
+                os.unlink(self.lock_file_path)
+            self._owns_lock = False
